+++ motor/mcconf_default.h
@@ -47,6 +47,9 @@
 #ifndef MCCONF_L_IN_CURRENT_MIN
 #define MCCONF_L_IN_CURRENT_MIN			-60.0	// Input current limit in Amperes (Lower)
 #endif
+#ifndef MCCONF_L_IN_CURRENT_MAP_START
+#define MCCONF_L_IN_CURRENT_MAP_START	1.0		// Input current to Q axis current limit map start
+#endif
 #ifndef MCCONF_L_MAX_ABS_CURRENT
 #define MCCONF_L_MAX_ABS_CURRENT		130.0	// The maximum absolute current above which a fault is generated
 #endif